Controller clean up
